Fixing response behavior on api catches

+++ src/pages/Order/All/index.js
@@ -39,7 +39,7 @@ import {
 export default function AllOrders({ userId, companyInfo }) {
 	//	Order state variables
 	const [orders, setOrders] = useState([]);
-	const [orderA, setOrderA] = useState({});
+	const [order, setOrder] = useState({});
 	const [userPasswordOnDelete, setUserPasswordOnDelete] = useState("");
 
 	//	Modal settings
@@ -63,7 +63,7 @@ export default function AllOrders({ userId, companyInfo }) {
 	}, [orders]);
 
 	useEffect(() => {
-		async function loadOrder() {
+		async function fetchData() {
 			await api.get("orderAll", {
 				headers : {
 					"x-access-token": userId
@@ -73,41 +73,53 @@ export default function AllOrders({ userId, companyInfo }) {
 					setOrders(response.data);
 					setupWebSocket();
 				}
+			}).catch((error) => {
+				setTitle("Erro!");
+				if(error.response.status === 400) {
+					setMessage(error.response.data);
+					setToastShow(true);
+				} else if(error.response.status === 404) {
+					setOrders([]);
+				} else if(error.response.status === 500) {
+					setMessage(error.message);
+					setToastShow(true);
+				} else {
+					setMessage("Algo deu errado :(");
+					setToastShow(true);
+				}
 			});
 
 			setIsLoading(false);
 		}
 
-		loadOrder();
+		fetchData();
 	}, [userId]);
-
-
-	async function handleSetOrder(event, order) {
-		event.preventDefault();
-
-		setOrderA(order);
-		setModalOrderListing(true);
-	}
 
 	async function handleDeliver(event, order) {
 		event.preventDefault();
 
 		var orderOK = false;
 
-		await api.put("/order/" + order._id, { status:true }, {
+		await api.put("/order/" + order._id, { status : true }, {
 			headers : {
 				"x-access-token": userId,
 			}
-		}).then(() => {
-			orderOK = true;
+		}).then((response) => {
+			if(response.status === 200) {
+				orderOK = true;
+			}
 		}).catch((error) => {
 			setTitle("Erro!");
-			if(error.response && typeof(error.response.data) !== "object") {
+			if(error.response.status === 400) {
 				setMessage(error.response.data);
-			} else {
+			} else if(error.response.status === 404) {
+				setMessage(error.response.data);
+			} else if(error.response.status === 500) {
 				setMessage(error.message);
+			} else {
+				setMessage("Algo deu errado :(");
 			}
-			setModalAlert(true);
+			setToastShow(true);
 		});
 
 		if(orderOK) {
@@ -131,23 +143,29 @@ export default function AllOrders({ userId, companyInfo }) {
 				data.push(cardsNewQtd);
 			}
 
-			await api.put("/userUpdateCard", { cardsNewQtd: data }, {
+			await api.put("/userUpdateCard", { cardsNewQtd : data }, {
 				headers : {
 					"x-access-token": userId,
 					"order-user-id": order.user._id
 				}
-			}).then(() => {
-				setTitle("Pedido enviado!");
-				setMessage("Alterações feitas com sucesso!");
-				setModalAlert(true);
+			}).then((response) => {
+				if(response.status === 200) {
+					setTitle("Pedido enviado!");
+					setMessage("Alterações feitas com sucesso!");
+					setModalAlert(true);
+				}
 			}).catch((error) => {
 				setTitle("Erro!");
-				if(error.response && typeof(error.response.data) !== "object") {
+				if(error.response.status === 400) {
 					setMessage(error.response.data);
-				} else {
+				} else if(error.response.status === 404) {
+					setMessage(error.response.data);
+				} else if(error.response.status === 500) {
 					setMessage(error.message);
+				} else {
+					setMessage("Algo deu errado :(");
 				}
-				setModalAlert(true);
+				setToastShow(true);
 			});
 		}
 	}
@@ -162,36 +180,42 @@ export default function AllOrders({ userId, companyInfo }) {
 
 	async function handleDeleteOrders(event) {
 		event.preventDefault();
-		setModalDeleteOrder(false);
 
 		await api.delete("order", {
 			headers : {
 				"x-access-token": userId,
 				password: userPasswordOnDelete
 			}
-		}).then(() => {
-			deleteAllSockets();
-			setTitle("Todos pedidos apagados!");
-			setMessage("Alterações feitas com sucesso!");
-			setModalAlert(true);
+		}).then((response) => {
+			if(response.status === 200) {
+				setModalDeleteOrder(false);
+				deleteAllSockets();
+				setTitle("Todos pedidos apagados!");
+				setMessage(response.data);
+				setModalAlert(true);
+			}
 		}).catch((error) => {
 			setTitle("Erro!");
-			if(error.response && typeof(error.response.data) !== "object") {
+			if(error.response.status === 400) {
 				setMessage(error.response.data);
-			} else {
+			} else if(error.response.status === 404) {
+				setMessage(error.response.data);
+			} else if(error.response.status === 500) {
 				setMessage(error.message);
+			} else {
+				setMessage("Algo deu errado :(");
 			}
 			setModalAlert(true);
 		});
 	}
 
 	return (
-		<div className="all-container p-0 w-100">
+		<div className="all-container p-0 w-100 h-100">
 			<Push toastShow={toastShow} setToastShow={setToastShow} title={title} message={message} />
 			{isLoading ?
 				<Container className="d-flex h-100">
 					<Spinner
-						className="my-5 mx-auto"
+						className="m-auto"
 						style={{width: "5rem", height: "5rem"}}
 						animation="grow"
 						variant="warning"
@@ -204,8 +228,8 @@ export default function AllOrders({ userId, companyInfo }) {
 					</h1>
 					<CardDeck className="mx-3">
 						<Row xs={1} sm={2} md={3} className="d-flex justify-content-around m-auto w-100">
-							{orders.map(order => (
-								<Col key={order._id} className="my-2">
+							{orders.map((orderI) => (
+								<Col key={orderI._id} className="my-2">
 									<Card text="white" bg="dark">
 										<Card.Header>
 											<Row>
@@ -213,20 +237,20 @@ export default function AllOrders({ userId, companyInfo }) {
 													<Image
 														className="w-100"
 														style={{ borderRadius: "50%" }}
-														src={order.user.thumbnail ? process.env.REACT_APP_API_URL + order.user.thumbnail_url: camera}
+														src={orderI.user.thumbnail ? process.env.REACT_APP_API_URL + orderI.user.thumbnail_url: camera}
 														alt="thumbnail"
 														fluid
 													/>
 												</Col>
 												<Col className="ml-3">
 													<Row>
-														<strong>{order.user.name ? order.user.name : null}</strong>
+														<strong>{orderI.user.name ? orderI.user.name : null}</strong>
 													</Row>
 													<Row>
-														<span>{order.user.email ? order.user.email : null}</span>
+														<span>{orderI.user.email ? orderI.user.email : null}</span>
 													</Row>
 													<Row>
-														<span>{order.creationDate ? order.creationDate : null}</span>
+														<span>{orderI.creationDate ? orderI.creationDate : null}</span>
 													</Row>
 												</Col>
 											</Row>
@@ -234,39 +258,39 @@ export default function AllOrders({ userId, companyInfo }) {
 										<Card.Body>
 											<Card.Text>
 												<p>
-													{order.phone ? "Telefone para contato: " + order.phone : "Telefone não informado"}
+													{orderI.phone ? "Telefone para contato: " + orderI.phone : "Telefone não informado"}
 												</p>
 												<p>
-													{order.deliver ?
-														"Endereço de entrega: " + order.address.join(", ")
+													{orderI.deliver ?
+														"Endereço de entrega: " + orderI.address.join(", ")
 														:
 														"Irá retirar no balcão!"
 													}
 												</p>
 												<p>
-													{order.deliver ?
+													{orderI.deliver ?
 														"Tempo para entrega: De " + companyInfo.timeDeliveryI + " a " + companyInfo.timeDeliveryF + " minutos"
 														:
 														"Tempo para retirada: " + companyInfo.timeWithdrawal + " minutos"
 													}
 												</p>
 												<p>
-													{"Total a pagar R$" + order.total}
+													{"Total a pagar R$" + orderI.total}
 												</p>
 												<p>
 														Método de pagamento:
-													{order.typePayment === 1 ?
+													{orderI.typePayment === 1 ?
 														" Cartão"
 														:
 														" Dinheiro"
 													}
 												</p>
 												<p>
-													{(order.change === order.total) ?
+													{(orderI.change === orderI.total) ?
 														"Não precisa de troco"
 														:
-														((order.typePayment === 0) ?
-															"Pagará R$" + order.change + ", troco de R$" + (order.change - order.total)
+														((orderI.typePayment === 0) ?
+															"Pagará R$" + orderI.change + ", troco de R$" + (orderI.change - orderI.total)
 															:
 															"Pagará na maquininha"
 														)
@@ -278,15 +302,15 @@ export default function AllOrders({ userId, companyInfo }) {
 													variant="light"
 													id="btn-custom-outline"
 													className="m-1"
-													onClick={e => handleSetOrder(e, order)}
+													onClick={() => { setOrder(orderI); setModalOrderListing(true); }}
 												>
 													Ver pedido
 												</Button>
-												{!order.status ?
+												{!orderI.status ?
 													<Button
 														variant="outline-warning"
 														className="m-1"
-														onClick={e => handleDeliver(e, order)}
+														onClick={e => handleDeliver(e, orderI)}
 													>
 														Entregar pedido
 													</Button>
@@ -322,7 +346,7 @@ export default function AllOrders({ userId, companyInfo }) {
 				centered
 			>
 				<Modal.Header closeButton>
-					<Modal.Title>Pedido de {orderA.user ? orderA.user.name : null }</Modal.Title>
+					<Modal.Title>Pedido de {order && order.user ? order.user.name : null }</Modal.Title>
 				</Modal.Header>
 				<Modal.Body>
 					{isLoading ?
@@ -333,7 +357,7 @@ export default function AllOrders({ userId, companyInfo }) {
 							variant="warning"
 						/>
 						:
-						<ProductDeck products={orderA.products} />
+						<ProductDeck products={order && order.products ? order.products : []} />
 					}
 				</Modal.Body>
 				<Modal.Footer>
